fix cifar10 tfrecords

+++ tensor2gan/data_generators/cifar10.py
@@ -25,10 +25,10 @@ def read_batch(filepath):
 
     images = np.array(data[b'data'], dtype=np.float32)
     # cifar 10 saved as channel first
-    images = np.reshape(images, [-1, 3, 32,32])
+    images = np.reshape(images, [-1, 3, 32, 32])
     images = images.transpose([0, 2, 3, 1])
 
-    labels = np.array(data[b'labels'], dtype=np.float32)
+    labels = np.array(data[b'labels'], dtype=np.int64)
     return images, labels
     
 def get_files(train=True, data_dir="./data"):
@@ -104,6 +104,7 @@ class GenerateCifar10(DataGenerator):
             dataset = dataset.map(utils.parse_record)
             dataset = dataset.map(parse_features)
             dataset = dataset.shuffle(batch_size * 5).batch(batch_size).repeat()
+            # dataset = dataset.batch(batch_size).repeat()
             iterator = dataset.make_one_shot_iterator()
             return iterator.get_next()
